pyproject.toml
[tool.poetry]
name = "discord-cms"
version = "0.1.0"
description = "Headless CMS API for Discord"
authors = ["Charles de Freitas <charles@defreitas.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
sqlalchemy = "^1.3.19"
discord-interactive = "^4.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
